Make sure to store number of child ranks, and correct size count for Double messages
We weren't correctly storing the number of ranks of a newly created child, which led to errors where the variable would go out of scope.

In addition, I fixed a problem where _getSize would return twice the correct size for Double messages. MPI_Probe returns the word count, not the array size. Because a Double is two words, we have to halve the result to get the true value. Otherwise, we end up sending the message and then an equal sized list of junk.